Changes on the pipeline creation process and execution (2)
