+++ repairnator/src/main/java/fr/inria/spirals/repairnator/process/inspectors/ProjectInspector4Bears.java
@@ -12,73 +12,56 @@ import fr.inria.spirals.repairnator.process.step.AbstractStep;
 import fr.inria.spirals.repairnator.process.step.BuildProject;
 import fr.inria.spirals.repairnator.process.step.CloneRepository;
 import fr.inria.spirals.repairnator.process.step.GatherTestInformation4Bears;
-import fr.inria.spirals.repairnator.process.step.InspectPreviousBuild;
 import fr.inria.spirals.repairnator.process.step.PushIncriminatedBuild;
 import fr.inria.spirals.repairnator.process.step.TestPreviousBuild;
 import fr.inria.spirals.repairnator.process.step.TestProject;
 import fr.inria.spirals.repairnator.serializer.AbstractDataSerializer;
 
+/**
+ * Created by fermadeiral.
+ */
 public class ProjectInspector4Bears extends ProjectInspector {
 	private final Logger logger = LoggerFactory.getLogger(ProjectInspector4Bears.class);
-	
-	private Build previousBuild;
-	private boolean previousBuildFlag;
-	
-	public ProjectInspector4Bears(Build build, String workspace, List<AbstractDataSerializer> serializers, String nopolSolverPath, boolean push, RepairMode mode) {
+
+	public ProjectInspector4Bears(Build build, String workspace, List<AbstractDataSerializer> serializers,
+			String nopolSolverPath, boolean push, RepairMode mode) {
 		super(build, workspace, serializers, null, push, mode);
 		this.previousBuildFlag = false;
 	}
-	
+
 	public void run() {
-        AbstractStep firstStep = null;
-        
-        AbstractStep cloneRepo = new CloneRepository(this);
-        AbstractStep buildRepo = new BuildProject(this);
-        AbstractStep testProject = new TestProject(this);
-        this.testInformations = new GatherTestInformation4Bears(this);
-        AbstractStep inspectPreviousBuild = new InspectPreviousBuild(this);
-        AbstractStep testPreviousBuild = new TestPreviousBuild(this);
+		AbstractStep firstStep = null;
+
+		AbstractStep cloneRepo = new CloneRepository(this);
+		AbstractStep buildRepo = new BuildProject(this);
+		AbstractStep testProject = new TestProject(this);
+		this.testInformations = new GatherTestInformation4Bears(this);
+		AbstractStep testPreviousBuild = new TestPreviousBuild(this);
 		AbstractStep buildRepoForPreviousBuild = new BuildProject(this);
-        AbstractStep testProjectForPreviousBuild = new TestProject(this);
-        AbstractStep gatherTestInformation = new GatherTestInformation4Bears(this);
-		
-        cloneRepo.setNextStep(buildRepo).setNextStep(testProject).setNextStep(this.testInformations)
-        	.setNextStep(inspectPreviousBuild).setNextStep(testPreviousBuild)
-        	.setNextStep(buildRepoForPreviousBuild).setNextStep(testProjectForPreviousBuild).setNextStep(gatherTestInformation);
-        
+		AbstractStep testProjectForPreviousBuild = new TestProject(this);
+		AbstractStep gatherTestInformation = new GatherTestInformation4Bears(this);
+
+		cloneRepo.setNextStep(buildRepo).setNextStep(testProject).setNextStep(this.testInformations)
+				.setNextStep(testPreviousBuild).setNextStep(buildRepoForPreviousBuild)
+				.setNextStep(testProjectForPreviousBuild).setNextStep(gatherTestInformation);
+
 		if (this.getPushMode()) {
 			PushIncriminatedBuild pushIncriminatedBuild = new PushIncriminatedBuild(this);
 			pushIncriminatedBuild.setRemoteRepoUrl(PushIncriminatedBuild.REMOTE_REPO_BEAR);
-            this.setPushBuild(pushIncriminatedBuild);
-            gatherTestInformation.setNextStep(pushIncriminatedBuild);
+			this.setPushBuild(pushIncriminatedBuild);
+			gatherTestInformation.setNextStep(pushIncriminatedBuild);
 		}
-        
-        firstStep = cloneRepo;
-        firstStep.setDataSerializer(this.serializers);
 
-        firstStep.setState(ProjectState.INIT);
+		firstStep = cloneRepo;
+		firstStep.setDataSerializer(this.serializers);
 
-        try {
-            firstStep.execute();
-        } catch (Exception e) {
-            this.addStepError("Unknown", e.getMessage());
-            this.logger.debug("Exception catch while executing steps: ",e);
-        }
-    }
-	
-	public Build getPreviousBuild() {
-		return this.previousBuild;
-	}
-	
-	public void setPreviousBuild(Build previousBuild) {
-		this.previousBuild = previousBuild;
-	}
-	
-	public boolean isAboutAPreviousBuild() {
-		return previousBuildFlag;
-	}
-	
-	public void setPreviousBuildFlag(boolean previousBuildFlag) {
-		this.previousBuildFlag = previousBuildFlag;
+		firstStep.setState(ProjectState.INIT);
+
+		try {
+			firstStep.execute();
+		} catch (Exception e) {
+			this.addStepError("Unknown", e.getMessage());
+			this.logger.debug("Exception catch while executing steps: ", e);
+		}
 	}
 }
